% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong.R
\name{convert_lat}
\alias{convert_lat}
\title{Internal main function for latitude conversion}
\usage{
convert_lat(str)
}
\arguments{
\item{str}{A string with a longitude value}
}
\description{
Internal main function for latitude conversion
}
