% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong.R
\name{convert_lon}
\alias{convert_lon}
\title{Internal main function for longitude conversion}
\usage{
convert_lon(str)
}
\arguments{
\item{str}{A string with a longitude value}
}
\description{
Internal main function for longitude conversion
}
