% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms-fxns.R
\name{dms}
\alias{dms}
\alias{pz_degree}
\alias{degrees}
\alias{minutes}
\alias{seconds}
\alias{pz_minute}
\alias{pz_second}
\title{extract degree, minutes, and seconds}
\usage{
pz_degree(lat = NULL, lon = NULL)

pz_minute(lat = NULL, lon = NULL)

pz_second(lat = NULL, lon = NULL)
}
\arguments{
\item{lat, lon}{(numeric/integer/character) one or more latitude or
longitude values. values are internally validated. only one of
lat or lon accepted.}
}
\value{
\code{pz_degree}: integer, \code{pz_minute}: integer, \code{pz_second}: numeric
}
\description{
extract degree, minutes, and seconds
}
