% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coords.R
\name{parse_lat}
\alias{parse_lat}
\title{parse latitude}
\usage{
parse_lat(lat, format = "")
}
\arguments{
\item{lat}{(character/numeric/integer) one or more latitude values}

\item{format}{format, deafult often works}
}
\value{
numeric vector of decimal degree coordinates
}
\description{
parse latitude
}
\examples{
parse_lat("45N54.2356")
parse_lat("-45.98739874")
parse_lat("N455698735", "HDDMMmmmmm")
parse_lat("40.123°")
parse_lat("40.123N74.123W")
parse_lat("N45 04.25764")
# bad values -> NaN
parse_lat("191.89")
# multiple
x <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
parse_lat(x)
system.time(parse_lat(rep(x, 10^4)))
}
