% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coords.R
\name{parse_lat_lon}
\alias{parse_lat_lon}
\title{parse latitude and longitude}
\usage{
parse_lat_lon(lat, lon, format = "")
}
\arguments{
\item{lat}{(character/numeric/integer) a latitude string/value}

\item{lon}{(character/numeric/integer) a longitude string/value}

\item{format}{format, deafult often works}
}
\description{
parse latitude and longitude
}
\examples{
parse_lat_lon(49.12, -120.43)
parse_lat_lon(93, -120.43)
parse_lat_lon(49.12, -190)
parse_lat_lon(92, -190)
}
