% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_llstr.R
\name{parse_latlon}
\alias{parse_latlon}
\title{Parse a character vector of latitude and longitude strings
This function parses geographic coordinate strings in various formats,
including compact cardinal forms (e.g., "N4.9196E101.347") and
different delimiters such as commas, spaces, semicolons, or dots.
It supports both "lat first" and "lon first" conventions.}
\usage{
parse_latlon(x)
}
\arguments{
\item{x}{(character) string with latitude and longitude, one or more in a vector.}
}
\value{
A data.frame with parsed latitude and longitude in decimal degrees.
}
\description{
Parse a character vector of latitude and longitude strings
This function parses geographic coordinate strings in various formats,
including compact cardinal forms (e.g., "N4.9196E101.347") and
different delimiters such as commas, spaces, semicolons, or dots.
It supports both "lat first" and "lon first" conventions.
}
\examples{
parse_latlon("N 04.1683, E 101.5823")
parse_latlon("N04.82344, E101.61320")
parse_latlon("N 04.25164, E 101.70695")
parse_latlon("N05.03062, E101.75172")
parse_latlon("N05.03062,E101.75172")
parse_latlon("N4.9196, E101.345")
parse_latlon("N4.9196, E101.346")
parse_latlon("N4.9196, E101.347")
parse_latlon("4.9196N, 101.345E")
# no comma
parse_latlon("N4.9196 E101.347")
parse_latlon("4.9196N 101.345E")
# no space
parse_latlon("N4.9196E101.347")
parse_latlon("4.9196N101.347E")

# DMS
parse_latlon("N4 51'36\", E101 34'7\"")
parse_latlon(c("4 51'36\"S, 101 34'7\"W", "N4 51'36\", E101 34'7\""))
# inverted
parse_latlon("E101.345, N04.9196")
parse_latlon("4.9196N 101.345E")
parse_latlon("4.9196N101.345E")
}
