% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_coords.R
\name{parse_lon}
\alias{parse_lon}
\title{parse longitude}
\usage{
parse_lon(lon, format = "")
}
\arguments{
\item{lon}{(character/numeric/integer) one or more longitude values}

\item{format}{format, deafult often works}
}
\value{
numeric vector of decimal degree coordinates
}
\description{
parse longitude
}
\examples{
parse_lon("45W54.2356")
parse_lon("-45.98739874")
parse_lon("N455698735", "HDDMMmmmmm")
parse_lon("40.123°")
parse_lon("40.123N74.123W")
parse_lon("N45 04.25764")
# bad values
parse_lon("181")
# many
x <- c("45W54.2356", "181", 45, 45.234234, "-45.98739874")
parse_lon(x)
system.time(parse_lon(rep(x, 10^4)))
}
