% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_parts.R
\name{parse_parts}
\alias{parse_parts}
\title{parse coordinates into degrees, minutes and seconds}
\usage{
parse_parts(str, format = "")
}
\arguments{
\item{str}{(character) string including latitude or longitude}

\item{format}{format, deafult often works}
}
\value{
data.frame with columns for:
\itemize{
\item degrees (integer)
\item minutes (numeric)
\item seconds (numeric)
}

NA given upon error
}
\description{
BEWARE: EXPERIMENTAL
}
\examples{
parse_parts("45N54.2356")
parse_parts("40.4183318")
parse_parts("-74.6411133")
parse_parts("-45.98739874")
parse_parts("40.123N")
parse_parts("40.123°")

# not working
parse_parts("N455698735")

# multiple
x <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
parse_parts(x)
# system.time(parse_parts(rep(x, 10^4)))
}
