% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pz_hemisphere.R
\name{pz_hemisphere}
\alias{pz_hemisphere}
\title{Determine Hemisphere Labels from Lat/Lon Strings}
\usage{
pz_hemisphere(lon, lat)
}
\arguments{
\item{lon}{A character vector of longitude strings}

\item{lat}{A character vector of latitude strings}
}
\value{
A character vector with hemisphere labels like "NW", "SE", etc.
}
\description{
Determine Hemisphere Labels from Lat/Lon Strings
}
\examples{
pz_hemisphere(c("45W", "170"), c("60N", "35S"))  # -> c("NW", "SE")
}
