% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pz_parse_lat}
\alias{pz_parse_lat}
\title{Parse vector of latitude strings to decimal degrees}
\usage{
pz_parse_lat(x)
}
\arguments{
\item{x}{A character vector of latitude strings}
}
\value{
A numeric vector of parsed decimal degrees
}
\description{
Parse vector of latitude strings to decimal degrees
}
