% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{pz_parse_lon}
\alias{pz_parse_lon}
\title{Parse vector of longitude strings to decimal degrees}
\usage{
pz_parse_lon(x)
}
\arguments{
\item{x}{A character vector of longitude strings}
}
\value{
A numeric vector of parsed decimal degrees
}
\description{
Parse vector of longitude strings to decimal degrees
}
