% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pz_parse_parts.R
\name{pz_parse_parts_lat}
\alias{pz_parse_parts_lat}
\title{Parse parts of latitude strings into degrees, minutes, and seconds}
\usage{
pz_parse_parts_lat(x)
}
\arguments{
\item{x}{A character vector of latitude strings}
}
\value{
A data frame with columns deg, min, and sec
}
\description{
Parse parts of latitude strings into degrees, minutes, and seconds
}
