% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pz_parse_parts.R
\name{pz_parse_parts_lon}
\alias{pz_parse_parts_lon}
\title{Parse parts of longitude strings into degrees, minutes, and seconds}
\usage{
pz_parse_parts_lon(x)
}
\arguments{
\item{x}{A character vector of longitude strings}
}
\value{
A data frame with columns deg, min, and sec
}
\description{
Parse parts of longitude strings into degrees, minutes, and seconds
}
