% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pz_parse_parts.R
\name{split_decimal_degree}
\alias{split_decimal_degree}
\title{Split decimal degree into degrees, minutes, and seconds}
\usage{
split_decimal_degree(x)
}
\arguments{
\item{x}{A numeric vector of decimal degrees}
}
\value{
A matrix with columns for degrees, minutes, and seconds
}
\description{
Split decimal degree into degrees, minutes, and seconds
}
