% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dms-fxns.R
\name{dms}
\alias{dms}
\alias{pz_degree}
\alias{pz_minute}
\alias{pz_second}
\alias{print.pz}
\alias{d}
\alias{m}
\alias{s}
\alias{+.pz}
\alias{-.pz}
\alias{/.pz}
\alias{*.pz}
\title{extract degree, minutes, and seconds}
\usage{
pz_degree(lat = NULL, lon = NULL)

pz_minute(lat = NULL, lon = NULL)

pz_second(lat = NULL, lon = NULL)

\method{print}{pz}(x, ...)

d(x)

m(x)

s(x)

\method{+}{pz}(e1, e2)

\method{-}{pz}(e1, e2)

\method{/}{pz}(e1, e2)

\method{*}{pz}(e1, e2)
}
\arguments{
\item{lat, lon}{(numeric/integer/character) one or more latitude or
longitude values. values are internally validated. only one of
lat or lon accepted}

\item{x}{(integer) an integer representing a degree, minute or second}

\item{...}{print dots}

\item{e1, e2}{objects of class pz, from using \code{d()}, \code{m()}, or \code{s()}}
}
\value{
\code{pz_degree}: integer, \code{pz_minute}: integer, \code{pz_second}: numeric,
\code{d}: numeric, \code{m}: numeric, \code{s}: numeric (adding/subtracting these also
gives numeric)
}
\description{
extract degree, minutes, and seconds
}
\details{
Mathematics operators are exported for \code{+}, \code{-}, \code{/}, and \code{*},
but \code{/} and \code{*} are only exported with a stop message to say it's not
supported; otherwise you'd be allow to divide degrees by minutes, leading
to nonsense.
}
\examples{
# extract parts of a coordinate value
pz_degree(45.23323)
pz_minute(45.23323)
pz_second(45.23323)

pz_degree(lon = 178.23423)
pz_minute(lon = 178.23423)
pz_second(lon = 178.23423)

pz_degree(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))
pz_minute(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))
pz_second(lat = c(45.23323, "40:25:6N", "40° 25´ 5.994 N"))

# invalid
pz_degree(445.23323)

# add or subtract
d(31)
m(44)
s(3)
d(31) + m(44)
d(31) - m(44)
d(31) + m(44) + s(59)
d(31) - m(44) + s(59)
d(-121) + m(1) + s(33)
unclass(d(31) + m(44) + s(59))
}
