% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hemisphere.R
\name{parse_hemisphere}
\alias{parse_hemisphere}
\title{parse coordinates into degrees, minutes and seconds}
\usage{
parse_hemisphere(lat, lon, format = "")
}
\arguments{
\item{lat}{(character/numeric/integer) one or more latitude values}

\item{lon}{(character/numeric/integer) one or more longitude values}

\item{format}{format, deafult often works}
}
\value{
character vector of quadrants, one of: NE, NW, SE, SW
}
\description{
BEWARE: EXPERIMENTAL
}
\examples{
# NE
parse_hemisphere("45N54.2356", "40.123N74.123W")
# NW
parse_hemisphere("40.4183318", "-120")
# SW
parse_hemisphere("-40.4183318", "-120")
# SE
parse_hemisphere("-40.4183318", "120")

# many inputs
library(randgeo)
pts <- rg_position(count = 1000)
lats <- as.character(vapply(pts, "[[", 1, 1))
lons <- as.character(vapply(pts, "[[", 1, 2))
parse_hemisphere(lats, lons)
}
