% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_lat_lon.R
\name{parse_lat_lon}
\alias{parse_lat_lon}
\title{parse latitude and longitude}
\usage{
parse_lat_lon(lon, lat)
}
\arguments{
\item{lon}{(character/numeric/integer) one or more longitude values}

\item{lat}{(character/numeric/integer) one or more latitude values}
}
\value{
data.frame, with columns lon, lat
}
\description{
parse latitude and longitude
}
\details{
length(lon) == length(lat)
}
\examples{
parse_lat_lon(-120.43, 49.12)
parse_lat_lon(-120.43, 93)
parse_lat_lon(-190, 49.12)
parse_lat_lon(240, 49.12)
parse_lat_lon(-190, 92)
# many
lons <- c("45W54.2356", "181", 45, 45.234234, "-45.98739874")
lats <- c("40.123°", "40.123N74.123W", "191.89", 12, "N45 04.25764")
parse_lat_lon(lons, lats)
}
